requirements-testing.txt
django-discover-runner==1.0
python-coveralls==2.2.0
